pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hit-modules"
version = "0.1.1"
description = "Shared middleware utilities for HIT microservices"
authors = [
    { name = "HIT Platform", email = "platform@hitlabs.dev" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-jose>=3.3.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
sqlalchemy = ["SQLAlchemy>=2.0.0"]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
]

[tool.setuptools]
packages = ["hit_modules"]
package-dir = { "" = "python" }

